refactor(api): PR-6.5 commit 5 — tag + captions + reg + version_config + training launch 23 routes
server.py 948 → 253 行（−695，−73%）。**最后一刀**：projects 域剩余 23 routes
全部抽到 training router，server.py 内 @app 装饰器降到 **0**（仅剩 import +
3 行 SPA mount + main re-export）。

  api/routers/projects/training.py    23 routes
                                       tag (1)
                                       captions read/write (3)
                                       captions snapshots create/list/restore/delete (4)
                                       captions commit / batch (2)
                                       reg preview-tags / get / build / caption / delete (5)
                                       reg_ai generate-prior + get task (2)
                                       version_config get/put/from_preset/save_as_preset (4)
                                       queue training launch (1)
                                       version_thumb (1)
  api/schemas/training.py             12 BaseModel  Wd14/CLTagger/LLMTagger Overrides +
                                                    TagJobRequest + CaptionEdit/CommitItem/
                                                    CommitRequest/BatchOp + RegBuildRequest +
                                                    RegAiRequest + FromPresetRequest +
                                                    SaveAsPresetRequest

行为零变更（fork_preset 用 _err_code = _preset_err_code 的别名导入，与
原 server.py 一致）；attention_backend `detect_attention_backend()` 走 late
import 保 reg_generate_prior 不动行为。

回归：PR-1 安全网 152 + project_endpoints 49 + tag_endpoints + reg_endpoints +
version_config + caption_snapshot = **273/273 全绿**。route snapshot 不变
（160 routes）。

**server.py 最终状态**：253 行，0 个 @app 装饰器，只剩 imports + lifespan/
middleware/router 通过 api/app.py 共享 + SPA mount + main re-export。
PR-7 起 server.py 完全变成空 shim 文件（仅保 `from .api.app import app`
给老 `from studio.server import app` 兼容）。
