+++ studio/api/app.py
@@ -38,6 +38,7 @@ from .routers.projects import crud as projects_crud
 from .routers.projects import exports as projects_exports
 from .routers.projects import curation as projects_curation
 from .routers.projects import ingestion as projects_ingestion
+from .routers.projects import training as projects_training
 from .routers.queue import io as queue_io_router
 from .routers.queue import lifecycle as queue_lifecycle
 from .routers.queue import outputs as queue_outputs
@@ -84,3 +85,5 @@ app.include_router(projects_exports.router)
 app.include_router(projects_ingestion.router)
 # PR-6.5 commit 4: files/thumb + curation + duplicates (12 routes)
 app.include_router(projects_curation.router)
+# PR-6.5 commit 5: tag + captions + reg + reg_ai + version_config + queue training + version_thumb (23 routes)
+app.include_router(projects_training.router)
